requirements.txt
click==8.1.3
config==0.5.1
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
fonttools==4.38.0
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
migrate==0.3.8
numpy==1.23.5
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
torch @ https://download.pytorch.org/whl/cu116/torch-1.13.0%2Bcu116-cp38-cp38-linux_x86_64.whl
torchaudio @ https://download.pytorch.org/whl/cu116/torchaudio-0.13.0%2Bcu116-cp38-cp38-linux_x86_64.whl
torchsummary==1.5.1
torchtext==0.14.0
torchvision @ https://download.pytorch.org/whl/cu116/torchvision-0.14.0%2Bcu116-cp38-cp38-linux_x86_64.whl
